Adjusting Timer Functionality
The changes accomplish the same thing as the old code with fewer lines. Calculates the difference of current hour with the start time, if not within the range of START/END time & sleeps for that amount as opposed to using the CHECK_INTERVAL_SECONDS which is used to sleep between tweet checks during active hours.